+++ lib/src/archive/compression_type.dart
@@ -1,2 +1,2 @@
 /// The format of the compression an [ArchiveFile] is stored with.
-enum CompressionType { none, deflate, bzip2 }
+enum CompressionType { none, deflate, bzip2, lzma }